Add `AbstractSoundSource` class
