+++ packages/dev/core/src/AudioV2/abstractAudio/abstractSound.ts
@@ -1,16 +1,14 @@
 import { Observable } from "../../Misc/observable";
 import type { Nullable } from "../../types";
 import { SoundState } from "../soundState";
-import { AbstractNamedAudioNode, AudioNodeType } from "./abstractAudioNode";
+import { AudioNodeType } from "./abstractAudioNode";
 import type { _AbstractSoundInstance } from "./abstractSoundInstance";
+import { AbstractSoundSource, type IAbstractSoundSourceOptions } from "./abstractSoundSource";
 import type { PrimaryAudioBus } from "./audioBus";
 import type { AudioEngineV2 } from "./audioEngineV2";
 import type { _AbstractAudioSubGraph } from "./subNodes/abstractAudioSubGraph";
 import type { IVolumeAudioOptions } from "./subNodes/volumeAudioSubNode";
 import { _GetVolumeAudioProperty, _GetVolumeAudioSubNode } from "./subNodes/volumeAudioSubNode";
-import type { AbstractAudioAnalyzer, IAudioAnalyzerOptions } from "./subProperties/abstractAudioAnalyzer";
-import type { AbstractSpatialAudio, ISpatialAudioOptions } from "./subProperties/abstractSpatialAudio";
-import type { AbstractStereoAudio, IStereoAudioOptions } from "./subProperties/abstractStereoAudio";
 import { _AudioAnalyzer } from "./subProperties/audioAnalyzer";
 
 /** @internal */
@@ -40,7 +38,7 @@ export interface IAbstractSoundPlayOptionsBase {
 /**
  * Options for creating a sound.
  */
-export interface IAbstractSoundOptions extends IAbstractSoundOptionsBase, IAbstractSoundPlayOptions, IAudioAnalyzerOptions, ISpatialAudioOptions, IStereoAudioOptions {
+export interface IAbstractSoundOptions extends IAbstractSoundOptionsBase, IAbstractSoundPlayOptions, IAbstractSoundSourceOptions {
     /**
      * The output bus for the sound. Defaults to `null`.
      * - If not set or `null`, the sound is automatically connected to the audio engine's default main bus.
@@ -63,16 +61,13 @@ export interface IAbstractSoundStoredOptions extends IAbstractSoundOptionsBase, 
 /**
  * Abstract class representing a sound in the audio engine.
  */
-export abstract class AbstractSound extends AbstractNamedAudioNode {
-    private _analyzer: Nullable<AbstractAudioAnalyzer> = null;
+export abstract class AbstractSound extends AbstractSoundSource {
     private _newestInstance: Nullable<_AbstractSoundInstance> = null;
-    private _outBus: Nullable<PrimaryAudioBus> = null;
     private _privateInstances = new Set<_AbstractSoundInstance>();
     private _state: SoundState = SoundState.Stopped;
 
     protected _instances: ReadonlySet<_AbstractSoundInstance> = this._privateInstances;
     protected abstract readonly _options: IAbstractSoundStoredOptions;
-    protected abstract _subGraph: _AbstractAudioSubGraph;
 
     /**
      * Observable for when the sound stops playing.
@@ -81,13 +76,6 @@ export abstract class AbstractSound extends AbstractNamedAudioNode {
 
     protected constructor(name: string, engine: AudioEngineV2) {
         super(name, engine, AudioNodeType.HAS_INPUTS_AND_OUTPUTS); // Inputs are for instances.
-    }
-
-    /**
-     * The analyzer features of the sound.
-     */
-    public get analyzer(): AbstractAudioAnalyzer {
-        return this._analyzer ?? (this._analyzer = new _AudioAnalyzer(this._subGraph));
     }
 
     /**
@@ -137,42 +125,6 @@ export abstract class AbstractSound extends AbstractNamedAudioNode {
     }
 
     /**
-     * The output bus for the sound. Defaults to `null`.
-     * - If not set or `null`, the sound is automatically connected to the audio engine's default main bus.
-     * @see {@link AudioEngineV2.defaultMainBus}
-     */
-    public get outBus(): Nullable<PrimaryAudioBus> {
-        return this._outBus;
-    }
-
-    public set outBus(outBus: Nullable<PrimaryAudioBus>) {
-        if (this._outBus === outBus) {
-            return;
-        }
-
-        if (this._outBus) {
-            this._outBus.onDisposeObservable.removeCallback(this._onOutBusDisposed);
-            if (!this._disconnect(this._outBus)) {
-                throw new Error("Disconnect failed");
-            }
-        }
-
-        this._outBus = outBus;
-
-        if (this._outBus) {
-            this._outBus.onDisposeObservable.add(this._onOutBusDisposed);
-            if (!this._connect(this._outBus)) {
-                throw new Error("Connect failed");
-            }
-        }
-    }
-
-    /**
-     * The spatial features of the sound.
-     */
-    public abstract spatial: AbstractSpatialAudio;
-
-    /**
      * The time within the sound buffer to start playing at, in seconds. Defaults to `0`.
      */
     public get startOffset(): number {
@@ -191,28 +143,6 @@ export abstract class AbstractSound extends AbstractNamedAudioNode {
     }
 
     /**
-     * The stereo features of the sound.
-     */
-    public abstract stereo: AbstractStereoAudio;
-
-    /**
-     * The output volume of the sound.
-     */
-    public get volume(): number {
-        return _GetVolumeAudioProperty(this._subGraph, "volume");
-    }
-
-    public set volume(value: number) {
-        // The volume subnode is created on initialization and should always exist.
-        const node = _GetVolumeAudioSubNode(this._subGraph);
-        if (!node) {
-            throw new Error("No volume subnode");
-        }
-
-        node.volume = value;
-    }
-
-    /**
      * Releases associated resources.
      */
     public override dispose(): void {
@@ -220,11 +150,7 @@ export abstract class AbstractSound extends AbstractNamedAudioNode {
 
         this.stop();
 
-        this._analyzer?.dispose();
-        this._analyzer = null;
-
         this._newestInstance = null;
-        this._outBus = null;
 
         this._privateInstances.clear();
         this.onEndedObservable.clear();
@@ -330,9 +256,5 @@ export abstract class AbstractSound extends AbstractNamedAudioNode {
             this._state = SoundState.Stopped;
             this.onEndedObservable.notifyObservers(this);
         }
-    };
-
-    private _onOutBusDisposed = () => {
-        this._outBus = null;
     };
 }
